backend/image-processing-lambda/requirements.txt
Pillow==10.1.0
boto3==1.34.0
botocore==1.34.0

